+++ src/components/MapsContent/MapsContent.styles.ts
@@ -6,6 +6,11 @@ export default createUseStyles((theme: ITheme) => ({
     padding: "0.5rem",
     flex: 1,
   },
+  button: {
+    fontSize: "1rem",
+    margin: "1rem 0",
+    width: "20%",
+  },
   mapNames: {
     display: "flex",
     justifyContent: "space-around",
@@ -17,6 +22,13 @@ export default createUseStyles((theme: ITheme) => ({
     cursor: "pointer",
     maxHeight: "4rem",
     borderRadius: "50rem",
+  },
+  playerIcon: {
+    height: "42px",
+    width: "42px",
+    "&:hover": {
+      cursor: "grab",
+    },
   },
   activeMap: {
     border: `0.25rem solid ${theme.buttonTextColor}`,